Done with the paper. Actually the paper uses exactly the same approach as me. For some reason my memory tricked.
What I wanted to do for hash consing is to flatten the node the wraps it in a value type. .NET does not do a good job of flattening it and probably boxes it under the hood which is why I got worse performance when I tried it.

What is going on here is that I forgot the old lessons and am now rethreading the same path. This is not always bad and is a good way to discover bugs in some cases.

Agh, this is such a pain. If a bunch of people looked over my code and told me I'd done a good job it would be a big blow. The reason for that would be because that would immediately set the ceiling on the abstraction the language is capable of. And since Spiral is as good as it gets, that would be bad news for the rest too.

Statically typed languages with extensional type systems are never going to be as expressive as Spiral. So that would only leave dynamic languages in terms of scaling.

Though I disagreed with Jdh on the issue of parallelization, he does have a point. He did not make the point explicitly, but it might come out of the holster at some time that for things to be scalable they need to be parallelizable.

...Let me try something.