Unfortunately changing the way pat_type_lit is evaluated makes absolutely no difference.
This is a surprise to me. I expected this node to be the biggest overhead, but it was nothing.

And even if I howl about data structures and hash consing, really there are only constant benefits to be gotten there. Immutable data structures are all balanced trees anyway.

I am guessing they all add up to the same thing in the end.

I am desperate here now.

Let me do one thing. I will change Expr so as to flatten the tags.

If that does not improve anything I'll really accept that I have done a good job of this.