So I am done for the day. Despite all my effort, maybe I speed up the compiler by 2.5%.
This is nothing to scoff at. If I knew in advance I'd speed it up this much, I'd gladly trade a day to get it.

The real reason why this bothers me is not the slowness per se.

It is the inverse reasoning.

Nothing is without cost.

Ultimately, compilation speed is proportional to the work done. And ultimately, compilation speed places the ceiling on abstraction possible.

Hypothetically, if there is some great abstraction that would need an hour of compilation - then absolutely it will be out of reach.

I'll get by with join points somehow, but the ceiling is not that far away from here. Maybe one or two orders of magnitude is the max a human would be filling to wait before trying a different approach.

I am actually confident that Spiral is not slow, but I do not want to accept it.

Since Spiral is not slow given it power and expressiveness, no other language will be able to beat this either.

This being the absolute reality brings everything into stark contrast and I do not want to go into the implications of this.

But I will just go and say it.

If the limit to abstraction in near, and the Singularity is near, then probably the program to trigger it is within an order of magnitude in size and complexity of what I have here in Spiral.

It stands to reason that if programming skill itself cannot scale the mountain of abstraction past a certain point then optimization will have to take over.

If anything of value can be drawn out, then it will have to be close to here. Processors won't get any faster and compilers are such a type of programs that the areas where improvement is happening will not affect them. Even if processors did get faster, they'd just be hamstrung by memory locality requirements.

It is really a vicious thing - the symbolic manipulation Spiral does is the worst case for the computer in any eventuality.

The LSTM cell taking 4s to compile a few days ago really did make me gaze into the abyss. Layout types and join points really saved me here.

Originally I introduced layout types to speed up renaming (which they failed to do in the parser example) but here they proved their worth as Spiral's core feature.

Most likely, Spiral is done. And the abyss is still there. The only choice then is to believe in myself and plunge deeper into chaos.

I admit that the reasoning here is the kind gamblers use 'I will definitely win so the ace will come out', but speculation even good one is always like that.

Whether something was the right reasoning or not is possible only in hindsight.

What is absolutely, certainly, 100% sure is that the ceiling is there and that I need to stop being a programmer because there is no future in it.

The future is in AI just as it has always been.