All, in all, probably this performance issue shaved a few days of my time. It is time to refocus.
I've been thinking some more. Yeah, it is true that maybe dynamic languages would scale in due to their easy compilation, but there is no way I can accept that unoptimized programs are the future.

It is interesting - optimization is the key. It is the magic that glues a fractured program together and linearizes it.

There is absolutely no way that unoptimization is the true path. Hence probably like in the 60s, the AI agents will have to write multiple programs concurrently and just wait the equivalent of days of for them to compile.

What the future Spiral will provide is good library support. Join points can be shared amongst concurrent processes, and Spiral will be structured like a caching interpreter.

Unlike now, it will have mechanisms for persistence - finishing compilation will not be cause for it to eject its memory.

Optimization in sense of machine learning might be the future of programming, but I can't imagine the classical stuff I am doing here will just go away as it is so powerful.

So there I have my answer.

If Spiral starts getting too slow, I will work on adding persistence to it.

With that done, let me focus on what will come. Those HRL talks...let me finally go through some of them. I liked Schmidhuber's talk, probably the others might be worth looking into.

When I am done, I will plug is softmax cross entropy and test it out. After that I will bring in sampling and check out whether the RNN is capable of generating sentences. That should be fun.

When that is done, I will focus on jacking up the speed and move to doing the tutorials. After that I will seriously start doing reinforcement learning on poker.

Seems like a swell plan.