10:35am. https://www.youtube.com/watch?v=e1MIC9r0DLY&index=7&list=PLfsVAYSMwskuaQ7fxEJaBZcZ2aXyc-5K-
I'll just read the paper.

Agh, for some reason the rest of the videos do not have slides. Combined with the grainy audio this makes it really hard to focus.

And two of the videos are missing. This is really the pits.

http://techtalks.tv/hrl-nips-2017/

Oh the talks work now.

http://techtalks.tv/hrl-nips-2017/?page=2

Actually it seems there are more than two missing. But since the talks are here now let me take a look.

Edit: Meant to push this, but at any rate I am done with this.

I am not sure what I expected, but this is basically just filler and I am wasting time with this.

I think I will forget about all the distraction from here on out and do some work on the library.

I covered all the ground.

Once more, let me see if I can just go in straight line to the goal. There is nothing really stopping me anymore.