```     inl error {fwd bck} label input s =         inl batch_size = primal input .span_outer |> to float         inl div_by_minibatch_size x = x / batch_size         inl cost,bck =             map_redo_map {                 fwd = {                     map_in = fwd                     redo = (+)                     neutral_elem = zero                     map_out = div_by_minibatch_size                     }                 bck = toa_map ((<<) div_by_minibatch_size) bck                 } (input, label) s         inl accuracy _ = accuracy label input s         {cost accuracy}, bck ```
This whole thing is rotten. If I want accuracy I should just do it all at once. Do not drag the `s` into it.

Actually since it struck me, I might as well go with the plan. The way I am doing it right now is a throwback to the old library.

This might be a good time to think about how I want to do non-differentiable layers.

And maybe I can add parallel layers while I am at it.

It never really felt right to me that the cost function should be responsible for accuracy.