I sort of went from focus to distraction just now so let me get back on track.
I've thought about it for a while, but just like last time, I've come to the same conclussion. There is no way to abstract those beyond the point they were at.

For the sample function I am going to have to start from scratch anyway.

Damn these loops are difficult and to think I am resorting to copy paste. But I guess it is for a good cause.

Well let me just do it. I'll plug in the kernels later. Let me make the sketch for the sample function. It won't be easy, but the time to do it has come.

I've thought about it enough.