I could have started 30m ago, but the only thing I felt then was boredom. Let me see if I can make a step forward now. I feel slightly more up for it.
Really, just fuck Cuda so much. It adds so much complexity to everything and has made made spend two years of my life trying to handle it. The complexity is not even in kernels themselves.

When this is over I pray that I can spend some time not giving a crap about it for a while.

Now...

The damn decoder.

I need an alternative reduction kernel for it.

Let me put some thought into it.