Let me start with this.
Yesterday it feels like I feel short of my goals, so for today I am going to restrict and focus on them as much as possible. Just do the decoder, encoder and the sampling function.

Just do that, and then the layers. Once I have that, I will be able to properly sample from the net and what comes next will be more fun than this.

I am really not that interested in char-RNN, and am now just doing this to complete the 1.5 year long challenge.

The pieces for this will be useful in what is to come. I am starving.