Softmax cross entropy does not work on the char RNN. I understand why too. Probably the weights are just too poorly initialized. Unlike with square, the log on the backwards step wrecks it.
If I want to get this to work, I need to add layer norm or maybe slip the outputs.

There is something I want to do.

Let me train it for 30 rounds so I can see if the error goes below 60 with both the standard and the LSTM RNN.

When I will do then is use the square error, but only take advantage of the softmax during the sampling phase.