1:10pm. https://github.com/denizyuret/Knet.jl#Benchmarks-1
Ah, yes this is the Julia framework that I remember.

https://github.com/denizyuret/Knet.jl/blob/master/examples/charlm/charlm.ipynb

1:15pm. https://github.com/karpathy/char-rnn/blob/master/model/LSTM.lua

```
  local top_h = outputs[#outputs]
  if dropout > 0 then top_h = nn.Dropout(dropout)(top_h) end
  local proj = nn.Linear(rnn_size, input_size)(top_h):annotate{name='decoder'}
  local logsoft = nn.LogSoftMax()(proj)
  table.insert(outputs, logsoft)
```

Let me try the linear decoder once more.

Also what is the learning rate he is using? For both it and the standard RNN?

Yeah, I am definitely using too large of a learning rate.

Let me go back to using a standard RNN except this time with a linear decoder and softmax cost function.

30 epochs should be enough to get it to converge.