Wow, this is not what I expected at all.
Unless it is some kind of kernel bug, which I really doubt, then this is really really hard to get going.

This wasn't what I expected at all.

Yeah, I can understand that Karpathy's setup is much better tuned. That much is clear to me, but I really want to make this work like so.

In poker I am going to get plenty of data where one of the fields is 0 for a long time. I can't have the net blowing up on that which means that cross entropy as a cost function is out.

You know, given the kind of welcome I got on the ML sub, how about I stop indulging my nonexistent readership and do it all optimally from the start?

Right now I am holding back so I can do the tutorials, but...

Forget those thoughts.

Forget the previous plan. Let me bring in layer norm and multiplicative integration.

I'll definitely need that for poker bots anyway. These difficulties I am experiencing here are actually an excellent lesson for me.

It is one thing to read those papers where they beat the SOTA, but that has nothing to do with me.

This right here is the true worth of NNs.

Let me bring out the bag of tricks.