``` u = sum a m = map (inl a -> a - u) a o'' = sqr m o' = sum o'' o = sqrt o' result = map (inl g, m -> g * m / o) (g, m) ```
```
dg, dm +=
    map (inl dresult, g, m ->
        inl x = dresult / o
        x * m, x * g
        ) (dresult, g, m)

do +=
    map_redo_map {
        map_in = inl dresult, g, m -> -dresult * g * m / (o * o)
        redo = (+)
        } (dresult, g, m)

da +=
    inl do' = do * o' ^ (-1/2) / 2
    inl f = (*) 2 * do'

    inl du = map_redo_map { redo=(+); map_out=negate } dm
    inl g = (+) du
    map (inl m,da -> f m + g da) (m,da)
```

Edit: This last part is really killing me.

I can't fit it in my head and it is like it keeps slipping away.

Somehow or the other, I ended up working on the derivatives for layer normalization for a few hours.

I am tired from the relentless focusing so I will call it a day here.

I am going to figure out how to pack these things tomorrow.

I need some more time to unravel it.

I am very nearly done.

Layer normalization is one of the things that was the major driver for me creating Spiral. It is not going to get away from me this time.