Yeah, it does train, just particularly poorly.
It is very slow at the start, but then does converge. I need to use a really small sequence. Increasing the batch size makes it go quite a bit more slowly.

But that is due to having to do more learning.

Also this is using standard RNNs. Highway LSTM just explodes with softmax cross entropy.

At this point I have to ask, is it really supposed to not have an activation at the end?