https://openreview.net/forum?id=rk5upnsxe Normalizing the Normalizers: Comparing and Extending Network Normalization Schemes
I am just sitting here thinking.

Yeah, at this point it is fairly obvious to me that the network is struggling greatly.

On a single short sequence with softmax cross entropy it converges just fine and then it gets progressively more harder for it as I increase the length along various dimensions.

Highway LSTM just explodes which makes me doubt its viability. Probably, an activation is needed before passing the hidden state to the next layer.

Square error does not work at all.

All in all, right now things work so unbelievably poorly that if I was trying to do this without prior knowledge, I'd just give up here.

All in all, I am actually amazed that they can get this to work so well.

At this point, I am not suspecting an implementation error.

And I have great expectations of layer norm and multiplicative integration.

Thanks to this experience though, in the future I am going to take a more skeptical view of the things I read in papers.

Let me read the paper on the top, have breakfast and then I will resume work on finishing the pseudocode for layer norm.

Then I will get to work on implementing it.