https://openreview.net/pdf?id=rk5upnsxe Normalizing The Normalizers- Comparing And Extending Network Normalization Schemes
https://arxiv.org/abs/1803.01271
An Empirical Evaluation of Generic Convolutional and Recurrent Networks for Sequence Modeling

The TCN paper is important. Maybe I will dump RNNs for temporal convolutions after all this is done.

Temporal convolutions are definitely a new thing for me after a long hiatus where I stopped following the field for over a year.

I intend to implement LN regardless, but in the LN paper it bothers me how just adding the standard LN did not help any on sequence tasks. So to start things off I am going to implement a peephole LSTM. I went off half cocked with that highway RNN it seems.

Also I am going to add gradient clipping. That thing would be trivial to add and maybe it will help. Though my understanding is that my difficulties here are separate from that. Really, I think I should be able to get this to work with standard SGD on a LSTM. I should not need to bring out the whole arsenal just for this. Do I really need Adam?

Before I resume I want to read the paper by He that introduced the temporal convolutions.

Also these latest results make it much more easier for me to accept the findings in the Snail paper.