It seems hard to believe, but the dataset is in fact transformed correctly. I had some extra time so I decided to do this instead of watching HxH.
I guess that only leaves Adam. Though it seems hard to believe, I am going to have to bring it out.

In literally every single paper they use tricks such as learning rate annealing and Adam and gradient clipping. I had really wanted to go without that, but now that it has come to this bring it on.

I implemented Adam long ago in late 2015 once before. It won't be hard to do again.

I guess if Adam fails as well, that would only leave the decoder. I would need to kick out the unused symbols and do it that way, but I really do not think it should come to that.

Adam has to work. It is just like with the autoencoders.

Though since it has come to this, why don't I go through the new course by Ng that deals specifically with sequence modeling.

It might give me a clue with regards to this.

Of course, I can always ask on the ML sub.

...No, actually I have not verified the dataset completely yet. What I have done just now is verified is that I did not violate array bounds.

I haven't verified that the way I am feeding the data is the correct sequence. For all I know, I might have accidentally permuted it.

I need to print it out.