But before I do that, I want to make a timer.
Edit: Done. It might be worth tracking down why sampler causes the code size to increase by 2k lines, a lot of those Cuda kernels, but let me stop here for a bit to catch my breath.

...Ah, actually I get it. The minibatch size. That has to be it.

Well, nevermind that then.

...Let me get back to the main subject.

I've done the timings. It takes it ~21s for a single iteration. Meaning a 1000 iterations would take a bit less than 6h.

No way am I waiting that long for this to train.

Now this is the spot where I should think about bringing in layer norm and multicative integration. And the spot where I should think about starting work on the optimizations in that Cuda post.

Getting the library to the point where I have the char RNN working is an absolutely remarkable achievement.

But if this was the only thing that Spiral could do, it would not be enough to separate itself from the rest of the pack.

Once I implement the optimizations in that blog post that will change decisively.

I will start with the most major ones which would be the wavefront iteration and fusing the pointwise operations.

I can't believe it that the time for this has come.

I've been working just to get to this point from late 2016.

There will be no turning back.

I'll checkpoint the repo so I can compare the improvements directly.