I was in a really good mood, so I decided to just take a nap for an hour and think about what comes next.
https://devblogs.nvidia.com/optimizing-recurrent-neural-networks-cudnn-5/

Really, the stuff here is trivial, I'll be done with it tomorrow or in two days.

So right now I am thinking about serialization. I really need a way of saving models, and it would only be right to extend that desire to serializing arbitrary Spiral types. That would be a worthy addition to my arsenal.

So, tomorrow I will implement the above optimizations, then after than I'll work on putting in layer norm. Then in addition to that I could do multiplicative integration as well. Since a RNN with that has a lot smaller footprint than a LSTM it should help me train the sequence model faster.

After that I will do serialization. I do not need if for the char RNN since I do not care about them, but for RL agents, I definitely will need it.

Also the idea that I had is that it might be good to swap static graphs for dynamic ones. That could be done using runtime tags and boxing which Spiral does support.

I'll keep that idea around for the future - unlike TF, Spiral is a lot more flexible innately and I do not need it right now.

Before I move on to RL, it might be worth looking into architectures with temporal convolution. If they train faster and easier, that could be a significant boon for me. But I should just put that work aside until I actually need it. I've delayed my entry long enough as it is. I need to start before I die of old age or starve on the street somewhere or end up homeless.

I'll start putting the pieces together tomorrow. I've done an enormous amount of build up just to get to this point, so it is only right that I get to savor it for a while.