https://arxiv.org/abs/1803.01271 An Empirical Evaluation of Generic Convolutional and Recurrent Networks for Sequence Modeling
I've linked to this before, but this paper really did inspire me. Yeah, this, Snail and the dilated RNN papers changed the way I look at NNs.

But now, I think I will finally move to making a MI architecture. That along with LN is what broke me in late 2016.

It is time I get that out of the way.

I will get my revenge.