This is a failure. Not so much the mi function, but the library itself is too complicated.
I am going to simplify it by removing those stupid nil checks.

For a library like Spiral's ML one, the most important thing is that it be fast and extensible.

I knew there was a problem when I myself take 3h to implement what should have otherwise take 1h only.

In fact it is more like 5h at this point. I spent most of this time refreshing my damn memory of what is what.

What an embarassment.

The d2 function is not even correct.

In the vast majority of cases I do not even need to check for nil. I do not need the kernels to do all those compilcated checks.

I do not need to take care of the case with multiple outputs.

What I need to do is to make it as straightforward as possible.

I am not going to compete with TF here. Some new architecture is going to come along and it will just die on its own.

The same goes for PyTorch.

A program is as much as defined by its restrictions than what it can do.

Making the library too flexible will inevitably make it slow.

I am going to call it a day here.

I'll trim the library tomorrow and then complete multiplicative integration. Layer norm will come after that.

Actually in its current form, MI is quite well fused. The thing I am missing would be to stream gemm. That will come after layer norm.

What an annoyance.

To think I've been working on this for so long only to have it turn out that I've done a bad job.

I was not thinking properly. At this point in my programming career, I am definitely realizing that every piece of functionality comes with a cost attached in runtime, compile time, and in terms of code complexity and maintenance burden.

A part of learning how to program in Spiral is knowing when not to overuse its features.

I am going to restrict those generic operations in the Learning module much more severely. If I really need to check for whether variables are dual in key places, I can just do it there when I need to.

The rest can just be type errors.

I am going to do it. I will crush the complexities, finish the library and finally start doing RL. It won't be long now.

I am going to program like crazy from here.