learning9 works. This was remarkably painless so far. This was worth it either way as it did lead to me discovering an error in toa_fold3.
Let me see if I can get 10 and 11 to compile.

At this point I am really tired and raring to stop for the day. I am even drowsy after eating, but I will persevere. I'll keep it up for a bit longer.

If I can't do it today, I'll bring it to a close tomorrow.