Jesus, I am finally back. I am damn tired physically.
But given how poorly it trains, I am really dying to train in on a single sequence and seeing whether I can get the net to repeat it. No matter how I slice it the result I am getting is too strange. Is sampling not working properly? I also considered that some inputs might cause it to hurl into the abyss, but in that case, I'd expect it to perform well at the start. This is not the case here. I get only crap all the time.

Let me track this down.