Well, I can be sure that the thing is wrong somehow now, but I can't be assed to figure out what is wrong as I am too tired.
There is just one thing I've realized though.

It is definitely time for me to put in some kind of sort function into the standard library. Right now it is pretty much impossible for me to figure out the way it is ordered as everything is so messy.

My hunch is that it is optimizing correctly, and just the sampling is wrong. I'll figure out what the problem is tomorrow.

Well, maybe I'll put the sort routine into the Array module later. Right now I am just too hungry.