Let me stop slacking. I got up early after a great night's sleep and it feels like I did jack today.
It is time to finally do the wavefront iteration. Even it is not completely sound, but it should be fine in the context of just the ML library. Whatever I am missing by not streaming the `gemm`s, I'll get back by having a lot of depth and using the wavefront.

After I am done, I'll modify the streams so they they include the event in the same structure.

Well, let me do it. It should not be too hard.

The whole streams thing just threw me so much off my game today. I did not realize that Cuda's concurrency model is pretty much completely broken though I should have had an inkling of it by now. I guess I needed to try it once before coming to a firm conclusion.

Really, by now all signs point that I should fuse the kernels in their entirety, but I am not ready for that right now. I need a GPU with cooperative groups for that.

Plus I will need to write my own gemm kernel as well. It is not the time for that just yet.

My plan is move towards removing the dependencies on Cuda libraries entirely, but things like writing my own gemms will have to wait.

They are simply not that important.

What is important that I make those dilated, layer normed RNNs and set them loose on poker. There will be time to attain perfection a few years down the road. Financial independence comes first.

So, let me just focus on the wavefront iteration for now.