There was in fact a data race lurking in there.
My hunch was correct.

With this, I am done with this phase. I am satisfied.

I need to plan out how to implement layer norm. I did not exactly get whether the paper got rid of the gain and the bias factors and I need to review that.

I also need once again take the derivative of it. I could do it using AD, but like with softmax, I'd bet there is a way do it all in one blow.

I need to focus on that.