Even with 4 layers, it seems in practice the difference is only 20%.
But that allow me to come to come to a conlussion. Since I would not be losing much by ditching the wavefront, what I should do is fully fuse the forward and backwards passes.

Actually, since I had no idea how much effective the wavefront iteration would be, I did not have the motivation to even consider this. I believed that the wavefront would be more effective.

Damn, I really need a new GPU, but it is going to have to earn itself.

I am nearly done for the day, so for the remainder I will try figuring out layer norm for tomorrow.

There is not much left until I am done with the library.

Edit: Forgot to push this an hour ago.