```             //map=inl (in,er) sum -> er * in * (1f32 - in) - in * (sum - er * in)             map=inl (in,er) sum -> (er - sum) * in ```
Let me try this new version. Though it does feel wrong somehow. Can it really work?

```
[|
    [|0.2925366; -0.718359; 0.09999694; -0.3931978|]
|]

[|
    [|0.5028772; -1.234876; 0.1718971; -0.6759161|]
|]
```

Oh, wow it can. This is exactly the correct result.

This is fantastic. I've done so much mental elaboration tonight and this proves that my method passes muster.

Now let me adjust the d1_seq_broadcast function as I will be using it for everything in layer norm. I managed to reduce it all to that function.