``` Starting timing for: iteration 0 The time was 00:00:05.9738421 for: iteration 0 Training: 212.706627957961 Starting timing for: iteration 1 The time was 00:00:05.9425372 for: iteration 1 Training: 214.095126656925 Starting timing for: iteration 2 The time was 00:00:05.9332750 for: iteration 2 Training: 213.62075177361 Starting timing for: iteration 3 The time was 00:00:05.9011630 for: iteration 3 Training: 213.225436042337 Starting timing for: iteration 4 The time was 00:00:05.9245731 for: iteration 4 Training: 212.725390658659 ```
As expected, it does not work.

To be honest, I'd be shocked if it did. These kernels are rather complex and there is plenty of room for failure.

Let me plug it in the feedforward layer and I'll print out what comes out of LN in the first layer. That should tell me something. After that I should put it through the gradient checker.