12:05pm. ``` r=-0.575103 r=6.331953 r=0.740446 r=4.110253 r=-15.979129 r=1.447166 r=1.367027 r=3.138509 r=0.542423 r=-1.123546 ```
These are the adjoints that I think should be correct.

Let me do the alternative version.

```
r=-0.575103
r=6.331953
r=0.740445
r=4.110253
r=-15.979128
r=1.447166
r=1.367027
r=3.138509
r=0.542423
r=-1.123546
```

12:10pm. The two versions are identical. The reason for that is that the norm is just replicated across the inner dimension so in this case it makes no difference. The tiny difference that I do get is due to numerical instability.

12:15pm.

```
[|
    [|0.5449323; 0.7150191; 0.5773278; 0.6603092; 0.1656073; -4.996212; 0.5927575; 0.6363801; 0.5724515; 0.5314269|]
|]
```

Let me assume that this is correct and the grad checking failures are due to its finickiness. If so, then fusing the operations should not lead to changes.

Let me give it a shot.

I'll branch once more.