I am thinking here. This is a shock to me, but I guess there is a reason why LN is not widely used.
I can think of two reasons why:

1) The authors are pulling numbers out of a hat.
2) My implementation is wrong.

And I am not suspecting the second point anymore.

Nonetheless, there are some intriguing things in that paper in that they managed to train raw RNN with Relus.

I am thinking that it might be worth it to implement Selu activation.

Also I am interesting in implementing batch norm. I know that it works so it might be useful for testing whether it works better than LN on the tasks I am interested in.

Actually, I am satisfied with assuming LN does not work.

I can just let it hang. I am not interesting in implementing BN after all. That thing is a dead end even if it works great.

I feel like moving on. Since this is a loss, I'll try mi with relus and the selus next.

I'll put my faith in Selus and dilated connections next.