Ok, I am done. How about I call it a day here? LN just continues to relentlessly underwhelm.
This is the opposite of BN.

I am focusing on the smoothing term like crazy, but I do not think it has anything to do with anything.

Let me bring the relu and I will try something else tomorrow. I'll extend the kernel I am suing and try both making the smoothing term a vector and putting in the affine layer - a proper one and not just the matrix multiply like I have up to now.