Tanh actually works much better than sigmoid in a standard MIRNN. The best of both is definitely LN + Relu.
Yeah, I suspected it would work better than sigmoid because layer norm centers around the mean. Sigmoid would not mix well with that, but tanh and relu do.

LN does not works as well BN did for me on Mnist, but almost 94% in the first iteration is more than good enough for me. Maybe it would be better with a wider net?

I think this is good enough for today.

Tomorrow I will fuse activation functions into the LN kernel and speed up a little that way.

But these results are good enough for me. LN+Relu will be my activation of choice going forward. Mi with it trains a lot better than a LSTM in fast.

I guess, I could say that everything went as as planned.