"9:35am. I got up two hours ago, but I think I got plenty of rest regardless. I've done a lot of thinking during that time. Now that I've completed layer norm, it feels like a lid has been opened and ideas are surging forth. I've figured out how to extend the d1_broadcast kernel so that it does replication in any dimension. For today, my aim will be to extend the kernel and fully fuse the linear part, layer norm and the activation. I won't bother making the activation generic since I have no reason to use anything other than relu.
Also I've been thinking how to do matrix multiplication and have come up with a plan. I've even thought about pooling and convolution. And there is even a level beyond that which I do not understand in the Winograd and FF transforms. I know it was a problem making Winograd generic for those guys. Maybe I'd have more luck with Spiral?

But I do not need convolution or pooling (for divisive normalization.)

Right now without a doubt, I have something really powerful.

By completing multiplicative integration and layer norm, I've fulfilled what directly drove me to create Spiral 1.5 years ago. The first part of the grand quest is truly complete. The next part would be to create a powerful agent of my design to serve as the expression of my desire.

Besides that, I need to do a lot of cleaning up. By all means, the `CudaKernel` module is like only half complete even only considering the 2d case. Eventually, I'd want to put in everything. But, I can pay as I go along.

I am going to do a change of design in the ML library. I am actually going to take biases out of matrix multiplication.

I think it would be fine if n^3 operations like them are separate from the rest. Biases should be a part of activation. This is without a doubt correct. It just goes to show how much I am still affected by my previous experience of making a ML library in F#.

Here is what I realized today - in order to generalize replicate all I need is `mapi`. That is it. Since in layer norm I cannot do a reduction without writing to global memory first, what I will do instead is use atomics. Easy as pie, no problem. The way to reduce replication is always addition so they are exactly what I need. Without a doubt, they would be a better choice than doing write and then a read from global memory. I'll blast the way with atomics.

The reason why I need `mapi` is simply because it would be too much of a pain to put in all the variations of replicate directly into the d1 broadcast kernel.

...Wow, the more I think about it, the better `mapi` seems. I could also put it in gemm's map out and add the biases that way.

Well, let me chill a little and then I will clean up those useless layers and fuse the LN with the activation function and the linear part in the MIRNN.

That will be my goal for the day."