``` fv=[(98, PrimT Float32T); (97, PrimT Float32T); (96, PrimT Float32T);  (95, PrimT Float32T);  (6, ArrayT (ArtCudaGlobal (PrimT UInt64T),PrimT Float32T));  (4, ArrayT (ArtCudaGlobal (PrimT UInt64T),PrimT Float32T));  (2, ArrayT (ArtCudaGlobal (PrimT UInt64T),PrimT Float32T)); (82, PrimT Int64T);  (0, ArrayT (ArtCudaGlobal (PrimT UInt64T),PrimT Float32T))] make_EnvStack4(var_98, var_97, var_96, var_95, var_6, var_4, var_2, var_82, var_0) ```
Let me wind up. I got distracted by Reddit for a bit because I needed to unwind.

Now what is EnvStack4 again?

```
__device__ __forceinline__ EnvStack4 make_EnvStack4(float * mem_0, long long int mem_1, float * mem_2, float * mem_3, float * mem_4, float mem_5, float mem_6, float mem_7, float mem_8){
```

Now that I look at it more closely, what is going on here is probably me printing arguments in the opposite order.

Yeah, it figures. It couldn't have been anything else. This error emerged by me doing testing this particular code path for the first time.

Edit:

"4:25pm.

```
The time was 00:00:04.3952954 for: iteration 0
Training: 175.505488676183
Starting timing for: iteration 1
The time was 00:00:04.4176466 for: iteration 1
Training: 133.892650464002
Starting timing for: iteration 2
The time was 00:00:04.3669848 for: iteration 2
Training: 123.777685614193
Starting timing for: iteration 3
The time was 00:00:04.3439931 for: iteration 3
Training: 117.618643844829
Starting timing for: iteration 4
The time was 00:00:04.3602114 for: iteration 4
Training: 113.612228113062
```

Wow, it is fast, but not correct unfortunately.

4:30pm. Let me commit this. I need to find where I made a mistake. The way I replicate is the obvious place to doubt."