The commits finally went through. Good.
I've been thinking what I want to do next and I think I will take the next week off from programming in order to update the manual. I think I have enough material now. I am also thinking that I will skip dilated RNNs and Reptile training for now. I had enough of the ML library for the time being. MI + LN is good enough to start. I will make the above pieces when I need them.

I am thinking that I should take what I have now and finally make the plunge into poker. I should get some sort of result out of it, and then I'll be able to tell whether adding layers and increasing the capacity of the net bring improvement. That will be the point to upgrade the architecture I am using.

This is good enough for now. I have a fully fused MI+LN+Relu net at my disposal and with it, I completely closed the gap with what I had in late 2016. I could work on the ML library for the next year, so instead of that it is time to move on the attack and make the first steps towards creating a usable agent.

So to mark the completion of a major milestone, I will work on the manual for a week or two.

It is time to start the fun part of my journey.